refactor(console): rename file
