+++ packages/console/src/utils/object.ts
@@ -1,0 +1,2 @@
+export const removeFalsyValues = (object: Record<string, unknown>) =>
+  Object.fromEntries(Object.entries(object).filter(([, value]) => value));
