+++ packages/console/src/utils/string.ts
@@ -1,4 +1,1 @@
 export const decapitalize = (value: string) => value.charAt(0).toLowerCase() + value.slice(1);
-
-export const removeFalsyValues = (object: Record<string, unknown>) =>
-  Object.fromEntries(Object.entries(object).filter(([, value]) => value));
